Crate revive_dev_runtime

Modules

api

currency

genesis_config_presets

Provides getters for genesis configuration presets.

Structs

CodeHashLockupDepositPercent

EthExtraImpl

Default extensions applied to Ethereum transactions.

ExistentialDeposit

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

TransactionByteFee

Version

Enums

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

Constants

RUNTIME_API_VERSIONS

VERSION

The runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Balance

Balance of an account.

Balances

Provides the ability to keep track of balances.

BalancesConfig

Block

Block type as expected by this runtime.

BlockNumber

An index to a block.

Hash

A hash of some data used by the chain.

Header

Opaque block header type.

Nonce

Index of a transaction in the chain.

OpaqueBlock

The opaque block type. This is the same BlockOf, but it has sp_runtime::OpaqueExtrinsic as its final extrinsic type.

Revive

Provides the ability to execute Smart Contracts.

ReviveConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

Sudo

Provides a way to execute privileged functions.

SudoConfig

System

Mandatory system pallet that should always be included in a FRAME runtime.

SystemConfig

Timestamp

Provides a way for consensus systems to set and check the onchain time.

TransactionPayment

Provides the ability to charge for extrinsic execution.

TransactionPaymentConfig

UncheckedExtrinsic